完成spring-activemq消息队列整合，实现新增商品，消费者同步监听

+++ dd-manager/dd-manager-interface/src/main/java/com/wdt/ddshop/service/ItemService.java
@@ -36,5 +36,5 @@ public interface ItemService {
      * @param content
      * @return
      */
-    int saveItem(TbItem tbItem, String content,String paramData);
+    Long saveItem(TbItem tbItem, String content,String paramData);
 }
